+++ apps/site/models/__init__.py
@@ -15,9 +15,8 @@ from localground.apps.site.models.lookups import StatusCode, UploadSource, \
 
 # overlays
 from localground.apps.site.models.marker import Marker
-from localground.apps.site.models.overlaysource import OverlaySource
-from localground.apps.site.models.overlaytype import OverlayType
-from localground.apps.site.models.tileset import TileSet
+from localground.apps.site.models.tileset import OverlaySource, \
+    OverlayType, TileSet
 
 # accounts
 # from localground.apps.site.models.base import Base
